fix(rueidisaside): order canceled fills on one pipeline
